requirements.txt
anthropic>=0.52.0,<2.0.0
resend>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
